backend/requirements.txt
customtkinter
python-nmap
scapy
shodan
packaging
